refactor: move string extension methods and tests to their own files
