+++ Testing/VelaptorTests/Content/AtlasJSONDataPathResolverTests.cs
@@ -10,8 +10,8 @@ using System.IO.Abstractions;
 using System.Reflection;
 using FluentAssertions;
 using Moq;
-using Velaptor;
 using Velaptor.Content;
+using Velaptor.ExtensionMethods;
 using Xunit;
 
 /// <summary>
