+++ Testing/VelaptorTests/Content/Fonts/FontTests.cs
@@ -14,12 +14,12 @@ using FluentAssertions;
 using Helpers;
 using Moq;
 using Newtonsoft.Json;
-using Velaptor;
 using Velaptor.Content;
 using Velaptor.Content.Caching;
 using Velaptor.Content.Exceptions;
 using Velaptor.Content.Fonts;
 using Velaptor.Content.Fonts.Services;
+using Velaptor.ExtensionMethods;
 using Velaptor.Graphics;
 using Velaptor.Services;
 using Xunit;
