+++ Testing/VelaptorTests/Content/SoundPathResolverTests.cs
@@ -10,8 +10,8 @@ using System.IO.Abstractions;
 using System.Reflection;
 using Helpers;
 using Moq;
-using Velaptor;
 using Velaptor.Content;
+using Velaptor.ExtensionMethods;
 using Xunit;
 
 /// <summary>
