+++ Velaptor/Content/Fonts/Font.cs
@@ -12,6 +12,7 @@ using System.Numerics;
 using System.Text;
 using Caching;
 using Exceptions;
+using ExtensionMethods;
 using Graphics;
 using Guards;
 using Services;
