+++ Velaptor/Services/AppSettingsService.cs
@@ -8,6 +8,7 @@ using System;
 using System.IO.Abstractions;
 using System.Text.Json;
 using Exceptions;
+using ExtensionMethods;
 using Guards;
 
 /// <inheritdoc/>
